Update README.md
This paper synthesizes research on technology innovations (AI, ML, etc.), its relation to the electronic commerce (EC) industry and their potential future effects on the industry. The future generation of EC or e-commerce is anticipated to use more digitally advanced technology for automation which may result in reduction of total cost, increased customer loyalty, satisfaction and higher return on investment. To be more effective and efficient in terms of customer’ satisfaction by being more profitable, fintech is to be considered among one of the fields that is to be expected more expanding and growing to meet the demands of this rapid digital transformation and play its role to automate the most integral parts of this industry by disrupting many of its business processes, structures, and operations. Modern EC is highly inclined toward utilizing technological advancements in the field of machine learning, augmented reality, deep learning, business intelligence, and artificial intelligence. In this brief research article, we review the literature on EC and forecast potential future industry heights with the upcoming arrival of innovative cutting-edge technical advancements as this sector might be anticipating in the future to achieve. Through this effort, investors may find innovative areas for investments that may yield attractive returns for their investments in the near future. The practitioners of EC industry could discover new directions for industrial advancements like the establishment of industry success key performance indicators, industry’s new research potential areas and technologies needs to be adopted as a career for this industry. Due to increased industry involvement in cutting-edge creative technologies, the regulators may also find some new areas that call for potential new controls to safeguard the security and ethics, as the sector develops under impending revolutionary technologies.